Move file from src/api to src/
