+++ src/plugin-api.js
